feat: Add hooks for redux dispatch and selector
This commit adds a new file hooks.ts that exports custom hooks useAppDispatch and useAppSelector for easier usage of eact-redux in the application. These hooks are used to access the dispatch function and select state from the Redux store respectively. The hooks are typed with the corresponding types AppDispatch and RootState from the store.ts file.