feat: Update ActionModal component to use new ActionTypeAll
This commit updates the ActionModal component to use the new ActionTypeAll type from the taskTypes.ts file. This allows the component to handle the edit action in addition to the existing delete and cancel actions. The ActionTypeAll type is now imported and used in the CardActions component as well. This change improves the functionality of the application by enabling users to edit tasks in addition to deleting and canceling them.